serial/tree_fs: safer `make_base_object()` wich checks archive also
Keep registry of archives that can be associated with object formatter.
`make_base_object()` extracts formatter and then also checks if current
archive is associated with this formatter.

That allows to prevent possible hidden behavior change if serialisation
to simple string happens while the same object is being saved to file.